// sdk/digitaltwins/digital-twins-parser-generator/src/parserGenerator/metamodelDigest.ts
export interface PropertyVersionDigest {
  idRequired: boolean;
  defaultLanguage?: string;
  allowed: boolean;
  class?: string;
  versions?: number[];
  maxCount?: number;
  minCount?: number;
  maxInclusive?: number;
  minInclusive?: number;
  maxLength?: number;
  pattern?: string;
  valueUris?: string[];
  values?: string[];
  typeRequired: boolean;
  uniqueProperties?: string[];
  value?: string | number | boolean;
  uniqueAmong: string[];
}

export interface MaterialPropertyDigest {
  _: {
    literal: boolean;
    class?: string;
    dictionaryKey?: string;
    abstract: boolean;
    datatype?: string;
    plural: boolean;
    optional: boolean;
    dtmiSegment?: string;
    inherited: boolean;
    shadowed: boolean;
    isKey: boolean;
    isSeg: boolean;
    description: string;
  };
  [dtdlVersion: number]: PropertyVersionDigest;
}